int to long id

+++ src/main/java/com/cars/a2/models/Utility.java
@@ -8,7 +8,7 @@ public class Utility {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String value;
 
@@ -20,11 +20,11 @@ public class Utility {
         this.value = value;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
